Remove TestModal, update UserDetailsModal
- Make UserDetailsModal use normal constructor
